requirements.txt
google-generativeai==0.7.2
faiss-cpu==1.7.4
numpy==1.26.4
streamlit==1.36.0
